Update content of KO translations
These have been reviewed by a native speaker.

Aside from grammar fixes the only content change is to increase the formality of the greeting:
"%{recipient}님 안녕하세요!" --> "%{recipient}님"

The old one was informal (Hello ___) and the new one is more formal which is safer for a more broad range of usage scenarios. Incidentally we do the same thing in Japanese with 様 (sama).