Handle capitalization of model name in error messages (#336)
Previously, we would always downcase the model name in the error partial. In some languages, this is not appropriate. For example, in German, nouns are always capitalized. In addition, the translated string might put the model name first, so it should still be capitalized.

There's not an elegant way to handle these kinds of rules, so we now define a helper devise_i18n_fix_model_name_case which capitalizes appropriately for the situation. The current implementation downcases in all cases, except in the error partial in German.

Fixes #260